requirements.txt
asgiref==3.7.2
Django==4.2.5
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.22
Pillow==10.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
